fix: improved on the suggestion
