+++ package.json
@@ -32,6 +32,7 @@
     "seed": "ts-node src/seed.ts"
   },
   "dependencies": {
+    "@css-inline/css-inline-linux-x64-gnu": "^0.14.1",
     "@css-inline/css-inline": "^0.14.1",
     "@faker-js/faker": "^8.4.1",
     "@nestjs-modules/mailer": "^2.0.2",
