+++ src/modules/email/dto/create-template-response.dto.ts
@@ -1,0 +1,13 @@
+import { HttpStatus } from '@nestjs/common';
+import { ApiProperty } from '@nestjs/swagger';
+
+export class CreateTemplateResponseDto {
+  @ApiProperty({ enum: HttpStatus, description: 'HTTP status code' })
+  status_code: HttpStatus;
+
+  @ApiProperty({ description: 'Response message' })
+  message: string;
+
+  @ApiProperty({ type: [String], description: 'List of validation errors', required: false })
+  validation_errors?: string[];
+}
